小工具/生成2D纹理数组纹理/requirements.txt
# 图片组转2D纹理数组纹理工具依赖包

# GUI框架
PySide6>=6.5.0

# 图像处理库
Pillow>=9.0.0

# 类型提示支持（Python 3.8+需要）
typing-extensions>=4.0.0
